requirements.txt
alembic==1.4.2
aniso8601==8.0.0
apispec==3.3.1
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-rest-api==0.17.1
Flask-RESTful==0.3.8
Flask-ShortUrl==0.2.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
psycopg2==2.8.5
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
webargs==6.1.0
Werkzeug==1.0.1
WTForms==2.3.1
